backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==2.1.0
supabase==2.3.4
pyjwt==2.8.0
djangorestframework-simplejwt==5.3.1
Pillow
django-filter==23.5
